Port to TC 7.15: downgrade to old ImGuiNET API generation, pin PunishLib submodule
TC's bundled Dalamud only has ImGui.NET/ImGuiScene, and PunishLib's current
commit + ECommons 3.2.0 are both firmly on the modern Dalamud.Bindings.ImGui
binding (net10.0-windows). Pinned the PunishLib submodule to its last
pre-API13 commit (see PunishLib's own commit for details) and ECommons to
3.0.0.8, then fixed the resulting API13+ drift in LazyLoot's own code:
- LazyLoot.cs: removed Dalamud.Game.Chat using (obsolete); IDtrBarEntry.OnClick
  is a plain Action in this API generation (no DtrInteractionEvent modifier-key
  info) - now always opens the config UI, losing ctrl-click/left-right-click
  cycling in this API generation; IChatGui.ChatMessage handler reverted to the
  old (XivChatType, int, ref SeString, ref SeString, ref bool) delegate instead
  of IHandleableChatMessage; TerritoryChanged handler is ushort, not uint
- ConfigUi.cs/Utils.cs: reverted Dalamud.Bindings.ImGui -> ImGuiNET;
  ImTextureID -> nint; IDalamudTextureWrap.Handle -> ImGuiHandle;
  ImU8String.Empty -> null for ImGui.Columns; ImGui.InputText needs an
  explicit uint cast for its maxLength parameter
- Roller.cs: Lumina.Excel.Sheets.ItemAction.Action.Value.RowId (lazy row) ->
  ItemAction.Type (plain ushort in this Lumina.Excel generation)
- LazyLoot.csproj: TargetFramework net10.0-windows7.0 -> net9.0-windows7.0
  (only .NET 9 stable SDK available), Dalamud.Bindings.ImGui HintPath ->
  ImGui.NET.dll

Builds clean with `dotnet build -c Release -p:DalamudLibPath=...`.

+++ LazyLoot/Utils.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
-using Dalamud.Bindings.ImGui;
+using ImGuiNET;
 using Dalamud.Interface.Utility;
 using ECommons.ImGuiMethods;
 using FFXIVClientStructs.FFXIV.Client.Game.UI;
@@ -115,7 +115,7 @@ internal static class Utils
             state.FocusOnOpen = false;
         }
         
-        ImGui.InputText("##popupListSearch", ref state.Query, inputMaxLength);
+        ImGui.InputText("##popupListSearch", ref state.Query, (uint)inputMaxLength);
         ImGui.Dummy(new Vector2(0, 4));
 
         var childSize = listSize ?? new Vector2(width, 200);
